.claude/worktrees/great-liskov-ed936c/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for laboratorio_v5"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<0.116.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "alembic>=1.13.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.0.1,<4.1",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "psycopg[binary]>=3.2.0",
  "anthropic>=0.39.0",
]

[build-system]

requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

